7th Commit
This commit is the final commit of the summer. THe application displays
a UI as a ListView. Each sensor/actuator is displayed in this listview.
Once clicked, the smart object will get the html document. An actuator
will turn on or off a device, e.g. a CD player. A sensor can read for
a device's current state, e.g. a thermometer.

A sensor will read